Move more metadata filter logic into actions
