+++ indico/modules/events/logs/client/js/containers/LogEntryModal.js
@@ -8,14 +8,7 @@
 import {connect} from 'react-redux';
 
 import LogEntryModal from '../components/LogEntryModal';
-import {
-  fetchLogEntries,
-  setDetailedView,
-  setMetadataQuery,
-  setPage,
-  viewNextEntry,
-  viewPrevEntry,
-} from '../actions';
+import {setDetailedView, showRelatedEntries, viewNextEntry, viewPrevEntry} from '../actions';
 
 const mapStateToProps = ({logs}) => ({
   currentViewIndex: logs.currentViewIndex,
@@ -33,11 +26,8 @@ const mapDispatchToProps = dispatch => ({
   nextEntry: () => {
     dispatch(viewNextEntry());
   },
-  relatedEntries: async metadataQuery => {
-    dispatch(setMetadataQuery(metadataQuery));
-    await dispatch(fetchLogEntries());
-    dispatch(setDetailedView(null));
-    dispatch(setPage(1));
+  relatedEntries: () => {
+    dispatch(showRelatedEntries());
   },
 });
